--- tests/diff/difference.csv
@@ -1,55 +1,0 @@
-Column0,Column1,Column2,Column3,Column4,Column5,Column6,Column7,Column8,Column9,Column10,Column11,Column12,Column13,Column14
-Article,Current,Marked,Marked,,,,,,,,,,,
-Store,Name,Barcode,Product,Name,Brand,Aisle,Department,Pack,Marked,at,Marked,At,Received,Approval
-Number,SR,by,As,,,,,,,,,,,
-Size,Time,Date,,,,,,,,,,,,
-Focal,FILIPPO,BERIO,,,,,,,,,,,,
-Michal,2024-10-28,2024-10-19,,,,,,,,,,,,
-SuperStore,107626417,BALSAMIC,VINEGAR,OF,GR18,Grocery,6,0,6,No,Stock,AUTO,,
-Jurkowski,15:45:44,23:00:00,,,,,,,,,,,,
-#101,MODENA,250ML,,,,,,,,,,,,
-08002210125070,,,,,,,,,,,,,,
-ITSU,SWEET,SOY,&,SEA,,,,,,,,,,
-Focal,,,,,,,,,,,,,,
-SALT,CRISPY,SEAWEED,Michal,2024-10-28,2024-10-21,,,,,,,,,
-SuperStore,113517532,GR18,Grocery,8,0,8,No,Stock,AUTO,,,,,
-THINS,MULTIPACK,4,X,Jurkowski,15:52:01,23:00:00,,,,,,,,
-#101,,,,,,,,,,,,,,
-05060960661395,5G,,,,,,,,,,,,,
-Focal,,,,,,,,,,,,,,
-Michal,2024-10-28,2024-10-19,,,,,,,,,,,,
-SuperStore,113510613,YUTAKA,KIMCHI,200G,GR18,Grocery,6,0,6,No,Stock,AUTO,,
-Jurkowski,15:51:27,23:00:00,,,,,,,,,,,,
-#101,,,,,,,,,,,,,,
-05014276701504,,,,,,,,,,,,,,
-Focal,,,,,,,,,,,,,,
-TOMATO,PUREE,WITH,Michal,2024-10-28,2024-10-19,,,,,,,,,
-SuperStore,106975674,GR19,Grocery,12,0,24,No,Stock,AUTO,,,,,
-GARLIC,135G,Jurkowski,15:32:26,23:00:00,,,,,,,,,,
-#101,,,,,,,,,,,,,,
-05010251614586,,,,,,,,,,,,,,
-Focal,,,,,,,,,,,,,,
-ILLY,ESPRESSO,COFFEE,Michal,2024-10-28,2024-10-19,,,,,,,,,
-SuperStore,100148150,GR21,Grocery,6,0,6,No,Stock,AUTO,,,,,
-250G,Jurkowski,12:03:49,23:00:00,,,,,,,,,,,
-#101,,,,,,,,,,,,,,
-08003753900438,,,,,,,,,,,,,,
-Focal,,,,,,,,,,,,,,
-OPTIONS,BELGIAN,CHOC,Michal,2024-10-28,2024-10-12,,,,,,,,,
-SuperStore,100138349,GR21,Grocery,30,0,30,No,Stock,AUTO,,,,,
-DRINK,11G,Jurkowski,12:08:19,23:00:00,,,,,,,,,,
-#101,,,,,,,,,,,,,,
-07612100054017,,,,,,,,,,,,,,
-Focal,STARBUCKS,BY,,,,,,,,,,,,
-Michal,2024-10-28,2024-10-21,,,,,,,,,,,,
-SuperStore,110778121,NESPRESSO,DARK,GR21,Grocery,8,-2,12,No,Stock,AUTO,,,
-Jurkowski,12:29:31,23:00:00,,,,,,,,,,,,
-#101,ESPRESO,10,X,57G,,,,,,,,,,
-07613037280364,,,,,,,,,,,,,,
-Focal,,,,,,,,,,,,,,
-HALLWAY,RUNNER,GREY,Home,&,Michal,2024-10-28,2024-10-09,,,,,,,
-SuperStore,112323402,LR04,6,5,5,No,Stock,AUTO,,,,,,
-50CM,X,150CM,Leisure,Jurkowski,11:34:31,23:00:00,,,,,,,,
-#101,,,,,,,,,,,,,,
-05010525188072,,,,,,,,,,,,,,
-Page,1,/,1,,,,,,,,,,,
